chore: refactor Beszel-related actions into their own module
