+++ src/actions/beszel/validator.ts
@@ -1,0 +1,5 @@
+import { z } from 'zod'
+
+export const installMonitoringToolsSchema = z.object({
+  serverId: z.string(),
+})
